+++ nodejs/aws-infra/experimental/elasticloadbalancingv2/listener.ts
@@ -18,8 +18,6 @@ import * as aws from "@pulumi/aws";
 import * as pulumi from "@pulumi/pulumi";
 
 import * as x from "..";
-import { Network } from "./../../network";
-
 import * as utils from "./../../utils";
 
 export abstract class Listener
