+++ nodejs/aws-infra/experimental/elasticloadbalancingv2/listenerRule.ts
@@ -18,8 +18,6 @@ import * as aws from "@pulumi/aws";
 import * as pulumi from "@pulumi/pulumi";
 
 import * as x from "..";
-import { Network } from "./../../network";
-
 import * as utils from "./../../utils";
 
 /**
